Move service interface resources to a separate class
