+++ mico-core/src/main/java/io/github/ust/mico/core/REST/ServiceController.java
@@ -28,10 +28,8 @@ public class ServiceController {
     public static final String PATH_VARIABLE_SHORT_NAME = "shortName";
     public static final String PATH_VARIABLE_VERSION = "version";
     public static final String PATH_VARIABLE_ID = "id";
-    private static final String PATH_VARIABLE_SERVICE_INTERFACE_NAME = "serviceInterfaceName";
-    private static final String PATH_PART_INTERFACES = "/interfaces";
-    private static final String PATH_DELETE_SHORT_NAME = "shortNameToDelete";
-    private static final String PATH_DELETE_VERSION = "versionToDelete";
+    public static final String PATH_DELETE_SHORT_NAME = "shortNameToDelete";
+    public static final String PATH_DELETE_VERSION = "versionToDelete";
 
     @Autowired
     private ServiceRepository serviceRepository;
@@ -61,26 +59,6 @@ public class ServiceController {
         return ResponseEntity.ok(
                 new Resources<>(serviceResources,
                         linkTo(methodOn(ServiceController.class).getVersionsOfService(shortName)).withSelfRel()));
-    }
-
-    //GET| /services/{shortName}/{version}/interface
-    @GetMapping("/{" + PATH_VARIABLE_SHORT_NAME + "}/{" + PATH_VARIABLE_VERSION + "}" + PATH_PART_INTERFACES)
-    public ResponseEntity<Resources<Resource<ServiceInterface>>> getInterfacesOfService(@PathVariable(PATH_VARIABLE_SHORT_NAME) String shortName,
-                                                                                        @PathVariable(PATH_VARIABLE_VERSION) String version) {
-        List<ServiceInterface> serviceInterfaces = serviceRepository.findInterfacesOfService(shortName, version);
-        List<Resource<ServiceInterface>> serviceInterfaceResources = serviceInterfaces.stream().map(
-                serviceInterface -> new Resource<>(serviceInterface, getServiceInterfaceLinks(serviceInterface, shortName, version))
-        ).collect(Collectors.toList());
-        return ResponseEntity.ok(new Resources<>(serviceInterfaceResources, linkTo(methodOn(ServiceController.class).getInterfacesOfService(shortName, version)).withSelfRel()));
-    }
-
-    @GetMapping("/{" + PATH_VARIABLE_SHORT_NAME + "}/{" + PATH_VARIABLE_VERSION + "}" + PATH_PART_INTERFACES + "/{" + PATH_VARIABLE_SERVICE_INTERFACE_NAME + "}")
-    public ResponseEntity<Resource<ServiceInterface>> getInterfaceByName(@PathVariable(PATH_VARIABLE_SHORT_NAME) String shortName,
-                                                                         @PathVariable(PATH_VARIABLE_VERSION) String version,
-                                                                         @PathVariable(PATH_VARIABLE_SERVICE_INTERFACE_NAME) String serviceInterfaceName) {
-        Optional<ServiceInterface> serviceInterfaceOptional = serviceRepository.findInterfaceOfServiceByName(serviceInterfaceName, shortName, version);
-        return serviceInterfaceOptional.map(serviceInterface ->
-                new Resource<>(serviceInterface, getServiceInterfaceLinks(serviceInterface, shortName, version))).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
     //TODO: Ambiguous endpoint with /services/shortName
@@ -333,13 +311,6 @@ public class ServiceController {
     private List<Resource<Service>> getServiceResourcesList(List<Service> services) {
         return services.stream().map(service -> new Resource<>(service, getServiceLinks(service)))
                 .collect(Collectors.toList());
-    }
-
-    private Iterable<Link> getServiceInterfaceLinks(ServiceInterface serviceInterface, String shortName, String version) {
-        LinkedList<Link> links = new LinkedList<>();
-        links.add(linkTo(methodOn(ServiceController.class).getInterfaceByName(shortName, version, serviceInterface.getServiceInterfaceName())).withSelfRel());
-        links.add(linkTo(methodOn(ServiceController.class).getInterfacesOfService(shortName, version)).withRel("interfaces"));
-        return links;
     }
 
     private Iterable<Link> getServiceLinks(Service service) {
